feat(#786): NAM keeps its Amp / Noise Gate / EQ tabs next to Capture
The first cut collapsed every engine control into a single Amp tab. The
engine's own split is the one worth keeping: Amp (input, output, A2 slim),
Noise Gate (toggle + threshold) and EQ (toggle + bass/mid/treble), with
Capture holding the manifest axes.

The engine tabs are declared with the specs themselves (`nam::params`),
since they are identical for every NAM — a plugin package or the generic
`neural_amp_modeler` loader — so no manifest ever repeats them; only the
capture axes are tagged by the manifest-aware layer (`block::nam_schema`).
`processor.rs` was over the 500-line cap, so the schema moved to the new
`nam::params` module and is re-exported from `nam::processor` unchanged.
